added /system_kills/ request & report, minor changes
