+++ geography_updater.py
@@ -1,4 +1,4 @@
-import requests
+from ash_esi_lib import request_data
 
 # flags
 UPDATE_REGIONS = True
@@ -6,16 +6,6 @@ UPDATE_CONSTELLATIONS = False
 UPDATE_SYSTEMS = False
 
 output_location = "csv/geography/"
-
-# requests data from EVE ESI for route
-# returns dictionary from json response
-# route example: "/regions/" or "/regions/10000003/"
-def request_data(route):
-    url_head = "https://esi.evetech.net/latest"
-    datasource = "?datasource=tranquility"
-    url = url_head + route + datasource
-    print("requesting " + route + "...")
-    return requests.get(url).json()
 
 print("hello captain :)")
